requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.31.0
stripe==7.5.0
payu-websdk==0.1.7
setuptools==68.2.2


